Removed showStepBundles
