+++ source/javascripts/components/unified-editor/WorkflowCard/components/StepListItem.tsx
@@ -1,18 +1,14 @@
+/* eslint-disable import/no-cycle */
 import { memo } from 'react';
 
 import StepService from '@/core/models/StepService';
-import useBitriseYmlStore from '@/hooks/useBitriseYmlStore';
 import useDefaultStepLibrary from '@/hooks/useDefaultStepLibrary';
 import StepBundleCard from '../../StepSelectorDrawer/components/StepBundleCard';
 
 import StepCard, { StepCardProps } from './StepCard';
 
 const StepListItem = (props: StepCardProps) => {
-  const { stepBundleId, stepIndex, workflowId = '', ...rest } = props;
-
-  const cvs = useBitriseYmlStore((s) => {
-    return Object.keys(s.yml.workflows?.[workflowId]?.steps?.[stepIndex] ?? {})[0];
-  });
+  const { cvs, stepBundleId, stepIndex, workflowId = '', ...rest } = props;
 
   const defaultStepLibrary = useDefaultStepLibrary();
 
@@ -26,7 +22,7 @@ const StepListItem = (props: StepCardProps) => {
     return <StepBundleCard isCollapsable cvs={cvs} stepIndex={stepIndex} workflowId={workflowId} {...rest} />;
   }
 
-  return <StepCard stepIndex={stepIndex} stepBundleId={stepBundleId} workflowId={workflowId} {...rest} />;
+  return <StepCard cvs={cvs} stepIndex={stepIndex} stepBundleId={stepBundleId} workflowId={workflowId} {...rest} />;
 };
 
 export default memo(StepListItem);
