+++ source/javascripts/pages/PipelinesPage/components/Drawers/Drawers.tsx
@@ -162,7 +162,6 @@ const Drawers = ({ children }: PropsWithChildren) => {
           onClose={closeDialog}
           onSelectStep={handleAddStep}
           onCloseComplete={unmountDialog}
-          showStepBundles={!stepBundleId}
         />
       )}
