+++ source/javascripts/pages/WorkflowsPage/components/Drawers/Drawers.tsx
@@ -155,7 +155,6 @@ const Drawers = ({ children }: PropsWithChildren) => {
           onClose={closeDialog}
           onSelectStep={handleAddStep}
           onCloseComplete={unmountDialog}
-          showStepBundles={!stepBundleId}
         />
       )}
