+++ spec/integration/test_bitrise.yml
@@ -313,7 +313,8 @@ workflows:
     steps:
     - git::https://github.com/bitrise-steplib/steps-script:
         title: A script step referenced by git URL
-        summary: All non-overwritten attributes should be the step's default value from step.yml in the step's repository.
+        summary: All non-overwritten attributes should be the step's default value
+          from step.yml in the step's repository.
         inputs:
         - content: |
             #!/usr/bin/env bash
@@ -479,3 +480,7 @@ step_bundles:
         - clone_depth: "0"
     - cache-pull@2.1.3: {}
     - deploy-to-bitrise-io@1.7.1: {}
+  test:
+    steps:
+    - activate-ssh-key@4.1.1: {}
+    - open-vpn@0.1.2: {}
